apps/linguistic/linguistic/requirements.txt
Django==4.1.4
djangorestframework
Markdown
django-filter